+++ keyoscacquire/docs/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+- v2.1.0:
+  - Added command line function list_visa_devices to list the addresses of the VISA instruments available
+  - Added command line fucntion path_of_config to show the path of config.py
+  -
+  - Changed dependency from visa to pyvisa (the package called visa on pypi is not pyvisa..!), and added tqdm dependency
 - v2.0.1: Cosmetic change in README, clarifying changelog for previous version
 - v2.0.0: When using set_options_getTrace_save(), channels are now comma separated in the csv to provide channel headings according to the data columns. This is not directly compatible with previous versions as these had two lines of preamble in csvs, whereas it is now three (Instrument info, columns descriptions, date and time). Also added a known issue to README
 - v1.1.1: Logging gives elapsed time in milliseconds, change in logging level for elapsed time